Important changes and improvements
* Removed @ prefix
* changed script calling convention to provide a safer alternative

Since the @ prefix was not exclusive, it is not suitable for voting and
similar tasks. Still, scripts might need to know who invoked them, while
being able to work standalone.
This is achieved with a change in the script's environment. The
environment variable NICKNAME is used to hold the caller's nickname,
future versions might add other IRC-specific data.